2024_ai_agents_in_langgraph/l3-agentic-search-tools/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
duckduckgo_search==7.2.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.29
langgraph==0.2.62
langgraph-checkpoint==2.0.9
langgraph-sdk==0.1.51
langsmith==0.2.10
lxml==5.3.0
msgpack==1.1.0
orjson==3.10.14
packaging==24.2
primp==0.10.0
pydantic==2.10.5
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
soupsieve==2.6
tavily-python==0.5.0
tenacity==9.0.0
tiktoken==0.8.0
typing_extensions==4.12.2
urllib3==2.3.0
